# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Careera API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface SignupRequest {
  fullName: string;
  email: string;
  password: string;
  collegeName: string;
  interests: string[];
  pastExperience?: string;
  referralCode?: string;
}

export type OrganizerSignupRequestRole =
  (typeof OrganizerSignupRequestRole)[keyof typeof OrganizerSignupRequestRole];

export const OrganizerSignupRequestRole = {
  President: "President",
  Member: "Member",
  Coordinator: "Coordinator",
} as const;

export interface OrganizerSignupRequest {
  name: string;
  email: string;
  password: string;
  organizationName: string;
  collegeName: string;
  location: string;
  role?: OrganizerSignupRequestRole;
}

export type CurrentUserRole =
  (typeof CurrentUserRole)[keyof typeof CurrentUserRole];

export const CurrentUserRole = {
  user: "user",
  organizer: "organizer",
  admin: "admin",
} as const;

export interface CurrentUser {
  id: number;
  fullName: string;
  email: string;
  role: CurrentUserRole;
  collegeName?: string;
  interests?: string[];
  isApproved?: boolean;
  referralCode?: string;
}

export interface AuthResponse {
  user: CurrentUser;
  message: string;
}

export interface UserProfile {
  id: number;
  fullName: string;
  email: string;
  collegeName?: string;
  interests?: string[];
  pastExperience?: string;
  referralCode?: string;
  referralCount?: number;
  createdAt?: string;
}

export interface UpdateProfileRequest {
  fullName?: string;
  collegeName?: string;
  interests?: string[];
  pastExperience?: string;
}

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus];

export const EventStatus = {
  upcoming: "upcoming",
  deadline_soon: "deadline_soon",
  closed: "closed",
  cancelled: "cancelled",
} as const;

export interface Event {
  id: number;
  name: string;
  description: string;
  organizerName: string;
  organizationName?: string;
  collegeName: string;
  location: string;
  fees?: number;
  registrationStartDate?: string;
  registrationDeadline: string;
  eventDate: string;
  tags?: string[];
  status: EventStatus;
  isRegistered?: boolean;
  registrationCount?: number;
  organizerId?: number;
  createdAt?: string;
}

export interface CreateEventRequest {
  name: string;
  description: string;
  collegeName: string;
  location: string;
  fees?: number;
  registrationStartDate?: string;
  registrationDeadline: string;
  eventDate: string;
  tags?: string[];
}

export interface Registration {
  id: number;
  eventId: number;
  userId: number;
  event?: Event;
  registeredAt: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  reminder: "reminder",
  cancellation: "cancellation",
  postponement: "postponement",
  announcement: "announcement",
  general: "general",
} as const;

export interface Notification {
  id: number;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export interface Announcement {
  id: number;
  title: string;
  message: string;
  createdAt: string;
}

export interface CreateAnnouncementRequest {
  title: string;
  message: string;
}

export interface AdminUser {
  id: number;
  fullName: string;
  email: string;
  collegeName?: string;
  interests?: string[];
  createdAt: string;
  lastLogin?: string;
}

export interface AdminOrganizer {
  id: number;
  name: string;
  email: string;
  organizationName: string;
  collegeName: string;
  location?: string;
  role?: string;
  isApproved: boolean;
  isRejected: boolean;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalOrganizers: number;
  pendingOrganizers: number;
  totalEvents: number;
  activeEvents: number;
  totalRegistrations: number;
}

export type ReferralInfoReferralsItem = {
  id: number;
  fullName: string;
  joinedAt: string;
};

export interface ReferralInfo {
  referralCode: string;
  referralCount: number;
  referrals: ReferralInfoReferralsItem[];
}

export type GetEventsParams = {
  filter?: GetEventsFilter;
};

export type GetEventsFilter =
  (typeof GetEventsFilter)[keyof typeof GetEventsFilter];

export const GetEventsFilter = {
  all: "all",
  upcoming: "upcoming",
  past: "past",
  recommended: "recommended",
} as const;