# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Careera API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: Get all events
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, upcoming, past, recommended]
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event (organizer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update event (organizer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete event (organizer only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Already registered or registration closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /organizers/announcements:
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
...
  /admin/organizers/{id}/approve:
    post:
      operationId: approveOrganizer
      tags: [admin]
      summary: Approve organizer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organizer approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/organizers/{id}/reject:
    post:
      operationId: rejectOrganizer
      tags: [admin]
      summary: Reject organizer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organizer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/announcements:
    post:
      operationId: adminCreateAnnouncement
      tags: [admin]
      summary: Create platform-wide announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "201":
          description: Announcement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Event:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        organizerName:
          type: string
        organizationName:
          type: string
        collegeName:
          type: string
        location:
          type: string
        fees:
          type: number
        registrationStartDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        eventDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [upcoming, deadline_soon, closed, cancelled]
        isRegistered:
          type: boolean
        registrationCount:
          type: integer
        organizerId:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - organizerName
        - collegeName
        - location
        - registrationDeadline
        - eventDate
        - status
    CreateEventRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        collegeName:
          type: string
        location:
          type: string
        fees:
          type: number
        registrationStartDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        eventDate:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - collegeName
        - location
        - registrationDeadline
        - eventDate
    Registration:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        event:
          $ref: "#/components/schemas/Event"
        registeredAt:
          type: string
          format: date-time
      required:
        - id
        - eventId
        - userId
        - registeredAt
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - createdAt
    CreateAnnouncementRequest:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
      required:
        - title
        - message
...