# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Careera API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  message: string;
  createdAt: Date;
}